% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_gpos}
\alias{as_gpos}
\title{Convert to GPos object}
\usage{
as_gpos(.data, ...)
}
\arguments{
\item{.data}{A character vector of genomic position strings}

\item{...}{Additional arguments (unused)}
}
\value{
A GPos object
}
\description{
Converts character strings representing single genomic positions
to GPos objects.
}
\examples{
as_gpos("chr1:1000")
as_gpos("chr1:1,000:+")
as_gpos(c("chr1:1000", "chr2:2000", "chr3:3000"))
}
