% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special-formats.R
\name{.handle_special_formats}
\alias{.handle_special_formats}
\title{Handle special genomic string formats}
\usage{
.handle_special_formats(x)
}
\arguments{
\item{x}{Character string}
}
\value{
Parsed genomic information
}
\description{
Handle special genomic string formats
}
