% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{GenomicCoordinates-coercion}
\alias{GenomicCoordinates-coercion}
\title{Coercion methods for genomic coordinates}
\description{
Enhanced coercion methods to convert character strings to GRanges, GPos,
and GInteractions objects with support for various string formats.
}
