% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{GenomicCoordinates-IRanges}
\alias{GenomicCoordinates-IRanges}
\title{Enhanced IRanges parsing}
\description{
Extensions to IRanges parsing to handle comma-separated numbers and
space-delimited coordinates.
}
