% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicCoordinates.R
\name{detect_genomic_class}
\alias{detect_genomic_class}
\title{Detect the appropriate class for genomic strings}
\usage{
detect_genomic_class(x)
}
\arguments{
\item{x}{Character string or vector}
}
\value{
Character vector of predicted classes
}
\description{
Utility function to determine what class a genomic string should be
parsed as, without actually performing the parsing.
}
\examples{
detect_genomic_class("chr1:1000-2000")
detect_genomic_class("chr1:1000")
detect_genomic_class(c("chr1:1-10|chr2:20-30", "1000-2000"))
}
